refactor: merge live Stripe tests into integration_stripe, runtime-gated
Remove integration-stripe-live feature flag. Live tests now live in
integration_stripe.rs and skip at runtime when STRIPE_SECRET_KEY is
not set (matching mppx pattern). CI passes STRIPE_SECRET_KEY as env
var to the test step.

Revert #[allow(dead_code)] on ResultExt — pre-existing, not ours.
