+++ .github/workflows/ci.yml
@@ -51,13 +51,11 @@ jobs:
       - uses: Swatinem/rust-cache@v2
       - run: cargo update -p native-tls
       - uses: taiki-e/install-action@cargo-hack
-      - run: cargo test --features tempo,stripe,server,client,axum,middleware,tower,utils,integration-stripe
-      - name: Live Stripe integration tests
-        if: env.STRIPE_SECRET_KEY != ''
-        run: cargo test --features integration-stripe-live --test integration_stripe_live
+      - name: Tests
+        run: cargo test --features tempo,stripe,server,client,axum,middleware,tower,utils,integration-stripe
         env:
           STRIPE_SECRET_KEY: ${{ secrets.STRIPE_SECRET_KEY }}
-      - run: cargo hack check --each-feature --no-dev-deps --skip integration,integration-stripe-live
+      - run: cargo hack check --each-feature --no-dev-deps --skip integration
       - name: Check examples
         run: find examples -name Cargo.toml -exec cargo check --manifest-path {} \;
